KSY-쿼리dsl_설정

+++ src/main/java/com/aico/aibayo/service/NotepadService.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Service
 public interface NotepadService {
-    List<NotepadEntity> getAllByKinderNo(int kinderNo);
+    List<NotepadEntity> getAllByKinderNo(Long kinderNo);
 }
